requirements.txt
black==23.1.0
click==8.1.3
colorama==0.4.6
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pydantic==1.10.5
PyYAML==6.0
typing_extensions==4.5.0
